third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-invalid-http-http-expected.txt

CONSOLE ERROR: The 'referrer' Content Security Policy directive has the invalid value "invalid". Valid values are "always", "default", "never", and "origin".
Test that invalid === never from HTTP to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is ""
PASS successfullyParsed is true

TEST COMPLETE
